Clean up SshManager UI code
