+++ packages/admin-ui/src/pages/system/components/Advanced/SshManager/ChangeStatus.tsx
@@ -1,26 +1,13 @@
 import React, { useState } from "react";
 import { api } from "api";
 import Button from "components/Button";
-import Card from "components/Card";
 import ErrorView from "components/ErrorView";
 import Input from "components/Input";
 import Ok from "components/Ok";
 import { ReqStatus, ShhStatus } from "types";
 import "./sshManager.scss";
 
-export function SshManager() {
-  return (
-    <Card spacing>
-      <div>Enable, disable SHH</div>
-      <SshStatusManager />
-
-      <div>Change the SSH port</div>
-      <SshPortManager />
-    </Card>
-  );
-}
-
-export function SshStatusManager() {
+export function SshManagerChangeStatus() {
   const [reqGetStatus, setReqGetStatus] = useState<ReqStatus<ShhStatus>>({});
   const [reqSetStatus, setReqSetStatus] = useState<ReqStatus<ShhStatus>>({});
 
@@ -88,46 +75,6 @@ export function SshStatusManager() {
       {reqSetStatus.error && (
         <ErrorView error={reqSetStatus.error} hideIcon red />
       )}
-    </>
-  );
-}
-
-export function SshPortManager() {
-  const [port, setPort] = useState("");
-  const [reqStatus, setReqStatus] = useState<ReqStatus>({});
-
-  async function updatePort() {
-    try {
-      setReqStatus({ loading: true });
-      await api.sshPortChange({ port: parseInt(port, 10) });
-      setReqStatus({ result: true });
-    } catch (e) {
-      setReqStatus({ error: e });
-      console.error("Error on sshPortChange", e);
-    }
-  }
-
-  return (
-    <>
-      <Input
-        placeholder="New SSH port i.e. 1024"
-        value={port}
-        onValueChange={setPort}
-        type="number"
-        append={
-          <Button
-            variant="dappnode"
-            disabled={!port || reqStatus.loading}
-            onClick={updatePort}
-          >
-            Change
-          </Button>
-        }
-      />
-
-      {reqStatus.loading && <Ok loading msg="Changing SSH port..."></Ok>}
-      {reqStatus.result && <Ok ok msg="Changed SSH port"></Ok>}
-      {reqStatus.error && <ErrorView error={reqStatus.error} hideIcon red />}
     </>
   );
 }
