+++ packages/admin-ui/src/pages/system/components/Advanced/SshManager/sshManager.scss
