+++ packages/admin-ui/src/pages/system/components/Advanced/index.tsx
@@ -1,4 +1,3 @@
-import Button from "components/Button";
 import SubTitle from "components/SubTitle";
 import React from "react";
 import { SshManager } from "./SshManager";
